requirements.txt
flask
Flask-HTTPAuth
functions-framework
google-cloud-dns
google-cloud-logging